refactor(animations): put transitions next to respective components
