+++ src/alert/alert-transition.ts
@@ -1,0 +1,5 @@
+import {NgbTransitionStartFn} from '../util/transition/ngbTransition';
+
+export const ngbAlertFadingTransition: NgbTransitionStartFn = ({classList}: HTMLElement) => {
+  classList.remove('show');
+};
